third_party/blink/web_tests/external/wpt/css/css-anchor-position/anchor-position-multicol-fixed-001-expected.txt

This is a testharness.js-based test.
[FAIL] .target 1
  assert_equals: \n<div class="target" data-offset-x="20" data-offset-y="20" data-expected-width="20" data-expected-height="30"></div>\nwidth expected 20 but got 40
[FAIL] .target 2
  assert_equals: \n<div class="target" data-offset-x="20" data-offset-y="40" data-expected-width="20" data-expected-height="30"></div>\nwidth expected 20 but got 130
Harness: the test ran to completion.
